добавляет OffersContainer
