+++ src/pages/main-page/main-page.tsx
@@ -1,30 +1,16 @@
-import { useCallback, useState, useEffect } from 'react';
-import cn from 'classnames';
-import PlaceCardList from '../../components/place-card-list/place-card-list';
-import Map from './../../components/map/map';
+import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { getCity, getFilterType, getIsOffersLoading, getSortedCards } from '../../store/offers-data/selectors';
+import { getCity, getIsOffersLoading, getSortedCards } from '../../store/offers-data/selectors';
 import CityList from '../../components/city-list/city-list';
-import Sorting from '../../components/sorting/sorting';
-import { createOfferLocations } from '../../helpers/create-offer-locations';
 import { Loader } from '../../components/loader/loader';
 import { fetchOffers } from '../../store/api-actions';
+import { OffersContainer } from '../../components/offers-container/offers-container';
 
 function MainPage(): JSX.Element {
-  const [activeCardId, setActiveCardId] = useState('');
   const activeCity = useAppSelector(getCity);
   const cards = useAppSelector(getSortedCards);
-  const filter = useAppSelector(getFilterType);
   const isLoading = useAppSelector(getIsOffersLoading);
   const dispatch = useAppDispatch();
-
-  const onMouseEnter = useCallback((id: string) => {
-    setActiveCardId(id);
-  }, []);
-
-  const onMouseLeave = useCallback(() => {
-    setActiveCardId('');
-  }, []);
 
   useEffect(() => {
     dispatch(fetchOffers());
@@ -37,15 +23,8 @@ function MainPage(): JSX.Element {
     );
   }
 
-  let locationForMap, offerLocations;
 
   const isCards = cards?.length !== 0;
-  if (isCards) {
-    locationForMap = cards[0].city.location;
-    offerLocations = createOfferLocations(cards);
-
-  }
-
 
   const noPlacesFound = (
     <div className="cities__places-container container">
@@ -55,38 +34,6 @@ function MainPage(): JSX.Element {
           {cards.length} places to stay in {activeCity}
         </b>
       </section>
-    </div>
-  );
-
-  const offersContainer = (
-    <div className="cities__places-container container">
-      <section className="cities__places places">
-        <h2 className="visually-hidden">Places</h2>
-        <b className="places__found">{cards.length} places to stay in {activeCity}</b>
-        <Sorting filter={filter}/>
-        <PlaceCardList
-          onMouseEnter={onMouseEnter}
-          onMouseLeave={onMouseLeave}
-          cards={cards}
-          className={cn(
-            'cities__places-list',
-            'places__list',
-            'tabs__content'
-          )}
-          cardBlockName='cities'
-        />
-      </section>
-      {
-        <div className="cities__right-section">
-          <section className="cities__map map">
-            <Map
-              selectedPointId={activeCardId}
-              locations={offerLocations}
-              mainLocation={locationForMap}
-            />
-          </section>
-        </div>
-      }
     </div>
   );
 
@@ -100,7 +47,7 @@ function MainPage(): JSX.Element {
         </section>
       </div>
       <div className="cities">
-        {isCards ? offersContainer : noPlacesFound}
+        {isCards ? <OffersContainer activeCity={activeCity} cards={cards}/> : noPlacesFound}
       </div>
     </main>
